// libs/ngx-masked-input/src/lib/interfaces/masked-input.interface.original.ts
export interface MaskedInputOptions {
  type: 'numeric';
  prefix: string;
  appendPrefix: boolean;
  suffix: string;
  prependSuffix: boolean;
  float: boolean;
  floatSeparator: string;
  precision: number;
  separateThousands: boolean;
  separator: string;
  leadingZero: boolean;
  enableEmpty: boolean;
}